+++ sdk/ml/azure-ai-ml/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Bugs Fixed
   - #39187 - Submitting a Command job created using the __call__ method on another Command is missing outputs.
   - #37738 - documentation fix for deployment entities.
+  - #38681 - documentation fix for pipeline decorator.
 
 ## 1.24.0 (2025-01-07)